Garden Assignment and Classes
Player is correctly assigned to their garden and a PlayerGarden object is created. The PlayerGarden object holds the player that is assigned and the gardens folders and allotments relative to that garden. A classes file was also added which holds the classes that will be used in this project within a module script.