Refactored planting and datahandling
Changed planting to be held all server side and visual events such as harvesting and pickup to be handled all client side through events. still need to implement server side check for harvest. After plant is ready for harvest, put in map variable and remove when harvested. need to write code to detmerine the rewards for harvesting.